// client/src/lib/validation.tsx
import * as z from "zod";

export const SignUpValidation = z.object({
    firstName: z.string().min(2, { message: 'too short'}),
    lastName: z.string().min(2, { message: 'too short'}),
    email: z.string().email(),
    password: z.string().min(7, { message: 'Password must be at least 7 characters'}),
    location: z.string().min(3, { message: 'is it real city?'}),
    picture: z.string()

  });

export const SignInValidation = z.object({
    email: z.string().email(),
    password: z.string().min(7, { message: 'Password must be at least 7 characters'}),

});

export const PostValidation = z.object({
  caption: z.string().min(5).max(2200),
  file: z.custom<File[]>(),
  location: z.string().min(2).max(100),
  tags: z.string(),
});

export const ProfileValidation = z.object({
  file: z.custom<File[]>(),
  name: z.string().min(2, { message: "Name must be at least 2 characters." }),
  username: z.string().min(2, { message: "Name must be at least 2 characters." }),
  email: z.string().email(),
  bio: z.string(),
});